Moved over the rotation menu to new system
Had to change how the values where stored in RAM from singed to unsigned to be able to use the generic menu features. Also had to add the MenuEntrySub as a parameter to the menu callback functions to be able fo display and store the correct values.

Did some major cleanup now that the special case of rotator menu is gone together with all uses of the old version of the sub menu struct.

+++ NuEVI/globals.h
@@ -37,8 +37,8 @@ extern unsigned short vibDirection; //direction of first vibrato wave UPWD or DN
 extern unsigned short fastPatch[7];
 extern byte rotatorOn;
 extern byte currentRotation;
-extern int rotations[4];
-extern int parallel; // semitones
+extern uint16_t rotations[4];
+extern uint16_t parallel; // semitones
 
 extern int touch_Thr;
